test : domain 생성에 따른 테스트 코드 추가 및 수정
