+++ src/test/java/step2/controller/LottoTicketMakerTest.java
@@ -5,7 +5,6 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
 
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 
@@ -25,26 +24,6 @@ public class LottoTicketMakerTest {
     public void 입력한_가격으로_로또_티켓_정상출력_갯수확인(int lottoAmount, int qty) {
         LottoTicketMaker lottoTicketMaker = LottoTicketMaker.of(lottoAmount);
         assertEquals(qty, lottoTicketMaker.getLottos().size());
-    }
-
-    @ParameterizedTest
-    @DisplayName("로또 구매 값 0인 경우 예외")
-    @ValueSource(ints = {0})
-    public void 로또_구매값_0원(int value) {
-        assertThatThrownBy(() -> LottoTicketMaker.purchaseAmountValidate(value))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("구입 금액 값이 없습니다.");
-
-    }
-
-    @ParameterizedTest
-    @DisplayName("로또 구매 값 1000원 이하인 경우 예외")
-    @ValueSource(ints = {100, 200, 500})
-    public void 로또_구매값_1000원이하(int value) {
-        assertThatThrownBy(() -> LottoTicketMaker.purchaseAmountValidate(value))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("구입 금액은 1000원부터 입력가능합니다.");
-
     }
 
 }
